GAME-36
Seperated shooting script from the player movement script and added shooting script methods into the weapon manager class, which seems reasonable, considering the fact that the reload function is placed in the same class respectively. Also modified missile and grenade sprites to have transparent background
