+++ BioDude/Assets/Scenes/initial.unity
@@ -1188,7 +1188,8 @@ Prefab:
       propertyPath: m_RootOrder
       value: 2
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 82165514026457094, guid: daf86fe2f2bb1e14d97bff61bd885bd0, type: 2}
   m_ParentPrefab: {fileID: 100100000, guid: daf86fe2f2bb1e14d97bff61bd885bd0, type: 2}
   m_IsPrefabParent: 0
 --- !u!1 &268895107 stripped
