+++ BioDude/Assets/Scripts/Player scripts/PlayerMovement.cs
@@ -5,8 +5,6 @@ using UnityEngine;
 public class PlayerMovement : MonoBehaviour
 {
     public float speed;
-    public GameObject pistolBullet;
-    public AudioSource pistolFire;
 
 
 
@@ -58,7 +56,7 @@ public class PlayerMovement : MonoBehaviour
         transform.rotation = Quaternion.Euler(new Vector3(0f, 0f, directionAngle * Mathf.Rad2Deg - 90));
         Debug.DrawLine(transform.position, transform.position + 10 * new Vector3(Mathf.Cos(directionAngle) , Mathf.Sin(directionAngle), 0), Color.blue);
         if (Input.GetMouseButtonDown(0))
-            Shooting();
+            GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().Shoot();
         if (Input.GetKeyDown(KeyCode.E))
             UseGrenade();
         if (Input.GetKeyDown(KeyCode.R) && GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().currentAmmo > 0 && GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().isReloading == false && GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().currentClipAmmo != GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().clipSize)
@@ -71,30 +69,5 @@ public class PlayerMovement : MonoBehaviour
         return directionAngle;
     }
 
-    private void Shooting()
-    {
-        //sometimes bullet spawns behind the player :D
-        if (GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().cooldownEnded && GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().activeWeapon != null && GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().isReloading == false)
-        {
-            if (GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().currentClipAmmo > 0)
-            {
-                pistolFire.Play();
-                float x = GameObject.FindGameObjectWithTag("PlayerWeaponSlot").transform.position.x;
-                float y = GameObject.FindGameObjectWithTag("PlayerWeaponSlot").transform.position.y;
-                float z = GameObject.FindGameObjectWithTag("PlayerWeaponSlot").transform.position.z;
-                GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().cooldownEnded = false;
-                GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().currentClipAmmo--;
-                StartCoroutine("Cooldown");
-                Instantiate(pistolBullet, new Vector3(x, y, z), transform.rotation);
-                if (GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().currentClipAmmo == 0 && GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().currentAmmo > 0)
-                    StartCoroutine(GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().Reload());
-            }
-        }
-    }
 
-    IEnumerator Cooldown()
-    {
-        yield return new WaitForSeconds(GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().activeWeapon.GetComponent<Weapon>().cooldown);
-        GameObject.FindGameObjectWithTag("PlayerWeaponSlot").GetComponent<WeaponManager>().cooldownEnded = true;
-    }
 }
