+++ BioDude/Assets/Scripts/PlayerShooting.cs
@@ -1,0 +1,24 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerShooting : WeaponManager {
+
+	// Use this for initialization
+	void Start () {
+		
+	}
+	
+	// Update is called once per frame
+	void Update () {
+		
+	}
+
+
+    public void Shooting()
+    {
+
+    }
+
+
+}
