+++ BioDude/Assets/Scripts/PlayerShooting.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 396bd84d62f230a448914572f1cff8ad
+timeCreated: 1525376714
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
